Move botocore specific helpers into botocore.py
